requirements.txt
aiogram==2.20
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
colorama==0.4.4
colorclass==2.2.2
config==0.5.1
decorator==4.4.2
distlib==0.3.4
Django==4.0.1
django-braces==1.15.0
django-ckeditor==6.2.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
django-model-utils==4.2.0
docopt==0.6.2
factories==1.2.0
factory-boy==3.2.1
Faker==12.3.0
filelock==3.4.2
frozenlist==1.3.0
future==0.18.2
huepy==1.2.1
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.3
moviepy==1.0.3
multidict==6.0.2
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
pip-upgrader==1.4.15
platformdirs==2.4.1
pluggy==1.0.0
proglog==0.1.9
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.6
PySocks==1.7.1
pytest==6.2.5
pytest-mock==3.7.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
responses==0.17.0
schedule==1.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==2.65.0
terminaltables==3.1.10
toml==0.10.2
tqdm==4.62.3
ujson==5.3.0
urllib3==1.26.8
virtualenv==20.13.0
yarl==1.7.2
